Moved movement into model
