+++ core/src/com/po/fuck/model/movement/BasicMovement.java
@@ -1,4 +1,4 @@
-package com.po.fuck.controller.movement;
+package com.po.fuck.model.movement;
 
 import com.badlogic.gdx.math.Vector2;
 import com.po.fuck.model.collision.Collidable;
